refactor(channel): share draft chunking resolver
