+++ extensions/telegram/src/bot.helpers.test.ts
@@ -64,46 +64,4 @@ describe("resolveTelegramDraftStreamingChunking", () => {
       breakPreference: "paragraph",
     });
   });
-
-  it("clamps to telegram.textChunkLimit", () => {
-    const cfg: OpenClawConfig = {
-      channels: { telegram: { allowFrom: ["*"], textChunkLimit: 150 } },
-    };
-    const chunking = resolveTelegramDraftStreamingChunking(cfg, "default");
-    expect(chunking).toEqual({
-      minChars: 150,
-      maxChars: 150,
-      breakPreference: "paragraph",
-    });
-  });
-
-  it("supports per-account overrides", () => {
-    const cfg: OpenClawConfig = {
-      channels: {
-        telegram: {
-          allowFrom: ["*"],
-          accounts: {
-            default: {
-              allowFrom: ["*"],
-              streaming: {
-                preview: {
-                  chunk: {
-                    minChars: 10,
-                    maxChars: 20,
-                    breakPreference: "sentence",
-                  },
-                },
-              },
-            },
-          },
-        },
-      },
-    };
-    const chunking = resolveTelegramDraftStreamingChunking(cfg, "default");
-    expect(chunking).toEqual({
-      minChars: 10,
-      maxChars: 20,
-      breakPreference: "sentence",
-    });
-  });
 });
